Move FeatureFlag to internal
