+++ node/src/main/kotlin/net/corda/node/services/config/NodeConfiguration.kt
@@ -4,7 +4,7 @@ import com.typesafe.config.Config
 import net.corda.common.configuration.parsing.internal.Configuration
 import net.corda.common.validation.internal.Validated
 import net.corda.core.context.AuthServiceId
-import net.corda.core.context.FeatureFlag
+import net.corda.core.internal.context.FeatureFlag
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.internal.TimedFlow
 import net.corda.core.internal.notary.NotaryServiceFlow
